+++ haka_mqtt/reactor.py
@@ -1088,7 +1088,7 @@ class Reactor(object):
                             chosen_postfix)
         elif family == socket.AF_INET6:
             ip6, port, flow_info, scope_id = sockaddr
-            self.__log.info("Found family=inet6 sock=%s proto=%s addr=%s:%d%s",
+            self.__log.info("Found family=inet6 sock=%s proto=%s addr=[%s]:%d%s",
                             socktype_str,
                             proto_str,
                             ip6,